0806-team-project/ltnqls11/requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
reportlab
plotly
numpy